+++ src/main/java/org/geowe/service/constraints/validator/DuplicateVertexValidator.java
@@ -13,35 +13,49 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
-package org.geowe.service.constraints;
+package org.geowe.service.constraints.validator;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
 
+import org.geowe.service.constraints.DuplicateVertex;
 import org.geowe.service.geometry.engine.JTSGeoEngineerHelper;
 
+import com.vividsolutions.jts.geom.GeometryFactory;
+import com.vividsolutions.jts.geom.Point;
+import com.vividsolutions.jts.operation.valid.RepeatedPointTester;
+
 /**
- * Tests whether this Geometry is simple. The SFS definition of simplicity
- * follows the general rule that a Geometry is simple if it has no points of
- * self-tangency, self-intersection or other anomalous points. Simplicity is
- *
- * See JTS geometry.isSimple javadoc
- *  
+ * Tests whether this Geometry has repeated vertex.
+ * 
  * @author rafa@geowe.org
  *
  */
-public class SimpleGeometryValidator implements ConstraintValidator<SimpleGeometry, String> {
+public class DuplicateVertexValidator implements ConstraintValidator<DuplicateVertex, String> {
 
 	private JTSGeoEngineerHelper helper;
+	private RepeatedPointTester repeatedPointTester;
 
 	@Override
-	public void initialize(SimpleGeometry constraintAnnotation) {
+	public void initialize(DuplicateVertex constraintAnnotation) {
 		helper = new JTSGeoEngineerHelper();
+		repeatedPointTester = new RepeatedPointTester();
+		
 	}
 
 	@Override
 	public boolean isValid(String wkt, ConstraintValidatorContext context) {
-		return helper.getGeom(wkt).isSimple();
+
+		boolean hasRepeatedPoint = repeatedPointTester.hasRepeatedPoint(helper.getGeom(wkt));
+		if (hasRepeatedPoint) {
+			context.disableDefaultConstraintViolation();
+			Point repeatedPoint = new GeometryFactory().createPoint(repeatedPointTester.getCoordinate());
+			context.buildConstraintViolationWithTemplate(
+					context.getDefaultConstraintMessageTemplate() + ": " + repeatedPoint.toText())
+					.addConstraintViolation();
+		}
+
+		return !hasRepeatedPoint;
 	}
 
 }